Added bells!
Hit a bell when an item is about to drop to reserve it (make it spawn at the bell's location).
Multiple bells can compete for the same item. The bell that received the most damage wins the contest.
If the contest ends in a tie the item will spawn at the default location.
Hittig a bell will shorten the drop timer's time left.

Added two bells to stage plain.