Added hitlag and knockback influence.
When an attack connects every entity involved receives some hitlag.
The amount of hitlag depends on the damage dealt and the base knockback of the attack.
This is done to highlight powerful attacks and to signal the hit to the players.
The damage is applied before the hitlag and the knockback at the end of it.

Added directional influence over the receiving knockback.
If a player receiving damage inputs a direction during the hitlag, they can influence the knockback received.
Tilting the stick in the same direction of the knockback can increase its amount to a maximum of x1.5.
Tilting the stick in the opposite direction of the knockback can reduce its amount to a minimum of x0.5.

Changes to clashing:
No longer cancels attacks.
The amount of hitlag caused by clashes depends on the damage of the attacks involved.