Changes to damage functions.
Now, damage functions take as parameter the attacking hitbox instead of the attack properties.
This allows the entities to access all the hitbox variables and functions.

Added variable weapon to Hitbox. Is the weapon that the hitbox belongs to. Defaults to the node owner.
Added functions get_user, get_knockback and get_hitlag to Hitbox.